# Adds key_min to the index_block