- `index_data_keys` now stores both the key_min and key_max.

- Negative lookup improvement, checking for key_min, avoiding loading the filter/data_block for keys not contained in the index block's key range.

Benchmark  for sequential IDs:

Before:
```
Sequential:

Benchmarking...
info: Benchmark running against { 127.0.0.1:44735 }
error(message_bus): error connecting to replica 0: error.ConnectionRefused
info(message_bus): connected to replica 0
1225 batches in 63.06 s
load offered = 1000000 tx/s
load accepted = 158575 tx/s
batch latency p00 = 0 ms
batch latency p10 = 14 ms
batch latency p20 = 18 ms
batch latency p30 = 20 ms
batch latency p40 = 20 ms
batch latency p50 = 22 ms
batch latency p60 = 24 ms
batch latency p70 = 24 ms
batch latency p80 = 25 ms
batch latency p90 = 26 ms
batch latency p100 = 2550 ms
transfer latency p00 = 0 ms
transfer latency p10 = 3455 ms
transfer latency p20 = 6328 ms
transfer latency p30 = 10891 ms
transfer latency p40 = 16313 ms
transfer latency p50 = 21801 ms
transfer latency p60 = 28159 ms
transfer latency p70 = 34521 ms
transfer latency p80 = 40832 ms
transfer latency p90 = 46979 ms
transfer latency p100 = 53074 ms
```

After:
```
Benchmarking...
info: Benchmark running against { 127.0.0.1:44159 }
info(message_bus): connected to replica 0
1226 batches in 60.78 s
load offered = 1000000 tx/s
load accepted = 164529 tx/s
batch latency p00 = 0 ms
batch latency p10 = 13 ms
batch latency p20 = 14 ms
batch latency p30 = 17 ms
batch latency p40 = 20 ms
batch latency p50 = 20 ms
batch latency p60 = 23 ms
batch latency p70 = 24 ms
batch latency p80 = 25 ms
batch latency p90 = 26 ms
batch latency p100 = 2598 ms
transfer latency p00 = 0 ms
transfer latency p10 = 3112 ms
transfer latency p20 = 6365 ms
transfer latency p30 = 11223 ms
transfer latency p40 = 16208 ms
transfer latency p50 = 21999 ms
transfer latency p60 = 28082 ms
transfer latency p70 = 33969 ms
transfer latency p80 = 39317 ms
transfer latency p90 = 45087 ms
transfer latency p100 = 50790 ms
```